Adds conflict score func
