+++ assistFunctions.h
@@ -1,5 +1,12 @@
+#ifndef AF_H_ /* Include guard */
+#define AF_H_
+
 int sum(int arr[], int n);
 
-int computeConflictScore(int N, int board[N]);
+int computeConflictScore(int N, int *board);
 
-int computeRandMutation(int N, int board[N]);
+int computeRandMutation(int N, int *board);
+
+void printBoard(int N, int *board);
+
+#endif
