+++ makefile
@@ -1,10 +1,7 @@
-all: program1 assistFunctions
+all: program1
 
 program1: main.c
-	mpicc -fopenmp -o main.x main.c
-
-assistFunctions: 
-	mpicc -fopenmp -o assistFunctions.x assistFunctions.c
+	mpicc -fopenmp -o main.x main.c assistFunctions.c
 
 clean:
 	rm *.x
